feat(MainLayout.test.tsx): add tests to cover different states and navigation functionality
